src/backend/app/requirements.consolidated.txt
fastapi>=0.100.0,<0.101.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
uvicorn>=0.23.0,<0.24.0
sqlalchemy[asyncio]>=1.4.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
aiohttp>=3.8.0
beautifulsoup4>=4.9.0
pytest>=6.0.0
pytest-asyncio>=0.14.0
pytest-cov>=2.10.0
httpx>=0.24.0
asyncpg>=0.27.0
email-validator>=2.0.0
bcrypt>=4.0.0
